Add files via upload
So, these are the assault mechs fixed. When possible, I fixed the assault mechs by removing patchwork materials, adding battle computer pieces, or increasing armor. One mech just needed ammo moved to other places to allow the new fixed slot system for stealth armos. But a few mechs needed their hardpoint distribution changed - missile to antipersonnel - in a few cases to support chaff countermeasures. And the BLR-1SHans could not have a warfare suite and a command console - I fixed that by replacing the warfare suite with a bloodhound probe and a guardian ECM. Also included: A fix for the pepperbox rifle, which was not being properly recognized as a rifle that uses medium rifle ammo for the purposes of mech engineer.